Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34746 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242343 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,746 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     <dbl>                                                 <POLYGON [m]>
#>  1 01001  15900000 ((851129.6 6563369, 849482.8 6564176, 849667.4 6565849, 8490…
#>  2 01002   9200000 ((888109.7 6545456, 888899.8 6547151, 889113.2 6549308, 8888…
#>  3 01004  24600000 ((886593 6540886, 886484.7 6542458, 886834.2 6544986, 885524…
#>  4 01005  15900000 ((850413.2 6544061, 850717.3 6546372, 848183 6547215, 847736…
#>  5 01006   5900000 ((899820.1 6520340, 900682.5 6518954, 901728 6519277, 903300…
#>  6 01007  33500000 ((887654.3 6545858, 887736.3 6547025, 886026.1 6548175, 8820…
#>  7 01008   5200000 ((881422.9 6541053, 878932.6 6540738, 879924.5 6538500, 8820…
#>  8 01009   6900000 ((905769.6 6521658, 906462.9 6521317, 906853.5 6523147, 9080…
#>  9 01010  29300000 ((918167.2 6541320, 917153.8 6541707, 914308.9 6541120, 9144…
#> 10 02180  17200000 ((743729.4 6952523, 741285 6951375, 740553.5 6951046, 739825…
#> # ℹ 34,736 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1230 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111329.5 ymin: 6049774 xmax: 1242343 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,230 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150700000 ((967694.8 6550219, 966610.9 6548763, 965649.6 6548244, …
#>  2 200000438 326800000 ((312908.5 6710510, 312603 6709369, 310698.9 6711370, 30…
#>  3 200000545 104800000 ((748496.8 6817507, 747471.6 6818763, 746799 6819020, 74…
#>  4 200000628 150100000 ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 83…
#>  5 200000800 338700000 ((629746 6718932, 631310.3 6717272, 631108 6716107, 6318…
#>  6 200000925 204400000 ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750, …
#>  7 200000933 970600000 ((658116.6 6688458, 657680.9 6686697, 654603.5 6686858, …
#>  8 200004802 401900000 ((1006943 6288260, 1008622 6287985, 1009847 6286962, 101…
#>  9 200005932 414900000 ((608056.9 6735353, 608691.2 6735824, 607263.9 6738882, …
#> 10 200005957 328500000 ((964840.4 6806088, 963976.2 6807368, 963012.5 6808171, …
#> # ℹ 1,220 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242343 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5762900000 ((835167.2 6539004, 834133.8 6539425, 833933.3 6540334, 835…
#>  2 02    7361500000 ((709032.4 6889686, 708350.3 6889711, 705854.9 6890250, 705…
#>  3 03    7340000000 ((666004.6 6563892, 664797.9 6564076, 663623.7 6565142, 663…
#>  4 04    6925800000 ((916159.9 6302361, 915178.5 6304594, 913802.1 6305681, 913…
#>  5 05    5548900000 ((916868.6 6348200, 915174 6346720, 913855.4 6347234, 91470…
#>  6 06    4298700000 ((1003358 6292416, 1003237 6294068, 1004275 6295611, 100345…
#>  7 07    5527700000 ((788033.7 6360072, 786566.1 6359438, 785617.5 6359557, 784…
#>  8 08    5228800000 ((790282.6 6920806, 790106.9 6921038, 788122 6921622, 78723…
#>  9 09    4890700000 ((546664.1 6185712, 545024.5 6187679, 543322.9 6188437, 542…
#> 10 10    6003300000 ((755024.3 6782295, 756156.1 6783594, 755421.1 6785503, 753…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242343 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12013200000 POLYGON ((601706.7 6835639, 600473.8 6837752, 600625.8 683…
#>  2 24    39151200000 POLYGON ((495492.5 6659013, 494845 6661224, 495737.2 66621…
#>  3 27    47787300000 POLYGON ((702445.6 6632753, 702985.9 6634400, 704513.2 663…
#>  4 28    29874700000 POLYGON ((366610.6 6835047, 366743.4 6835864, 365602.2 683…
#>  5 32    31806200000 POLYGON ((605920.9 6904796, 605623.4 6904413, 605433.3 690…
#>  6 44    57436400000 POLYGON ((758079.9 6780274, 757175.7 6781679, 756315.7 678…
#>  7 52    32081400000 MULTIPOLYGON (((331097 6610579, 330522.9 6614791, 329085.4…
#>  8 53    27206300000 MULTIPOLYGON (((151242.6 6775399, 151046.1 6776972, 150464…
#>  9 75    84035500000 MULTIPOLYGON (((343694.4 6243276, 343721.3 6245696, 344303…
#> 10 76    72727000000 MULTIPOLYGON (((455951.4 6182703, 454293.1 6181044, 452520…
#> 11 84    69710580000 POLYGON ((631177.9 6401429, 632628.1 6402926, 633013.2 640…
#> 12 93    31401500000 MULTIPOLYGON (((843019.1 6251688, 839589 6251765, 836290.8…
#> 13 94     8679500000 POLYGON ((1189769 6078685, 1190393 6079566, 1190644 608164…
#>