Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34746 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,746 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     [m^2]                                            <MULTIPOLYGON [m]>
#>  1 70177   9100000 (((921738.8 6760207, 921659.6 6760207, 921189.9 6760832, 922…
#>  2 89050   4000000 (((741563.4 6758626, 741320.1 6758008, 739273.9 6757967, 738…
#>  3 60414  10700000 (((660340 6905721, 660211.1 6905210, 659332 6905186, 658695.…
#>  4 80756   3700000 (((661403.5 7000774, 661018.1 7002027, 660347.7 7002280, 660…
#>  5 14478  10100000 (((507839.1 6886562, 509195.2 6884716, 510632.4 6884522, 510…
#>  6 27219   4400000 (((542995.8 6900422, 543112.4 6900476, 543708.9 6900377, 545…
#>  7 68149   2900000 (((1044950 6732060, 1044491 6730419, 1043757 6729491, 104300…
#>  8 30249   3400000 (((797259.9 6301672, 797651.4 6302046, 798774.4 6302539, 799…
#>  9 72363  29200000 (((516821.2 6779360, 515867.3 6781101, 514762.1 6780382, 514…
#> 10 27396   4200000 (((585658.9 6918990, 585542.5 6918351, 584672.7 6918058, 583…
#> # ℹ 34,736 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1230 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111282 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,230 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150700000 ((967194.8 6554664, 966748.4 6552733, 966712.4 6551457, …
#>  2 200000438 326800000 ((325279 6718117, 325297.8 6716916, 323594.8 6715722, 32…
#>  3 200000545 104800000 ((749128 6818118, 748496.8 6817507, 748719.7 6818137, 74…
#>  4 200000628 150100000 ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 83…
#>  5 200000800 338700000 ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 63…
#>  6 200000925 204400000 ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750, …
#>  7 200000933 970600000 ((649837.2 6683322, 649872.8 6684591, 648421.6 6686474, …
#>  8 200004802 401900000 ((994645.8 6278527, 994377.9 6279412, 993498.9 6280419, …
#>  9 200005932 414900000 ((608056.9 6735353, 608691.2 6735824, 607263.9 6738882, …
#> 10 200005957 328500000 ((963012.5 6808171, 962553.2 6807836, 961230.9 6808913, …
#> # ℹ 1,220 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5762900000 ((836447.9 6538772, 835167.2 6539004, 834133.8 6539425, 833…
#>  2 02    7361500000 ((712067.1 6888906, 710504.2 6889186, 709997.6 6889771, 709…
#>  3 03    7340000000 ((666004.6 6563892, 664797.9 6564077, 663623.7 6565142, 663…
#>  4 04    6925800000 ((913802.1 6305681, 913518.7 6306495, 912835.8 6306600, 911…
#>  5 05    5548900000 ((913795.6 6349572, 914132.6 6351611, 913402.6 6353234, 913…
#>  6 06    4298700000 ((1005824 6290060, 1004519 6291871, 1003358 6292416, 100323…
#>  7 07    5527700000 ((788462.6 6359716, 788033.7 6360072, 787160 6360093, 78656…
#>  8 08    5228800000 ((783106.5 6923376, 780978.2 6923085, 780320 6923064, 77879…
#>  9 09    4890700000 ((549585.3 6180563, 548401.5 6181779, 546936.7 6182805, 546…
#> 10 10    6003300000 ((756156.1 6783595, 755421.1 6785503, 753420.7 6786380, 753…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12013200000 POLYGON ((599697.6 6839200, 600054.9 6839870, 598443.7 683…
#>  2 24    39151200000 POLYGON ((495455 6660250, 494845 6661224, 495737.2 6662124…
#>  3 27    47787300000 POLYGON ((702445.6 6632753, 702985.9 6634400, 704513.2 663…
#>  4 28    29874700000 POLYGON ((366110 6838631, 366763.7 6839144, 366163.9 68405…
#>  5 32    31806200000 POLYGON ((605433.3 6905532, 605431.6 6906573, 606206.2 690…
#>  6 44    57436400000 POLYGON ((758919 6780000, 758079.9 6780274, 757175.7 67816…
#>  7 52    32081400000 MULTIPOLYGON (((331097 6610579, 330522.9 6614791, 329085.4…
#>  8 53    27206300000 MULTIPOLYGON (((150908.4 6769905, 149172.4 6769696, 148753…
#>  9 75    84035500000 MULTIPOLYGON (((341100.6 6237138, 340672.7 6237834, 341434…
#> 10 76    72727000000 MULTIPOLYGON (((455951.4 6182703, 454293.1 6181044, 452520…
#> 11 84    69710580000 POLYGON ((631741.8 6399699, 631084.9 6400391, 631177.9 640…
#> 12 93    31401500000 MULTIPOLYGON (((848031.6 6250180, 845329 6251305, 843019.1…
#> 13 94     8679500000 MULTIPOLYGON (((1182447 6085188, 1181095 6085529, 1181795 …
#>