Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34826 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 34,826 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     <dbl>                                            <MULTIPOLYGON [m]>
#>  1 43076  21890000 (((748166.1 6463827, 747512.2 6465082, 748162.4 6465513, 748…
#>  2 65462  16010000 (((455022.6 6263682, 454435.4 6264116, 453913.1 6262475, 453…
#>  3 29058  32760000 (((178183.7 6781582, 177738.5 6780262, 178285 6780064, 17834…
#>  4 22218  15540000 (((243953.6 6877260, 243491.9 6876618, 243054.9 6876860, 241…
#>  5 38250   4900000 (((889525.8 6504615, 889698.2 6504084, 888708.2 6503830, 888…
#>  6 86097  27220000 (((496845 6582010, 497689.1 6583549, 498147.7 6583726, 49855…
#>  7 27058  36530000 (((574623 6898578, 573804.1 6899465, 573913.6 6900680, 57364…
#>  8 42225   8670000 (((821353.7 6493133, 822119.5 6492921, 822431 6493626, 82175…
#>  9 65478   6520000 (((451686.2 6205143, 450702.1 6205205, 449116.3 6206205, 449…
#> 10 06043  17140000 (((1046530 6320331, 1046759 6319283, 1047343 6318288, 104736…
#> # … with 34,816 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111318.1 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150670000 ((966748.4 6552733, 966712.4 6551457, 967694.8 6550219, …
#>  2 200000438 326800000 ((319454 6714465, 320465.5 6714207, 321096.5 6713767, 32…
#>  3 200000545 104800000 ((750344.5 6818238, 749128 6818118, 748496.8 6817507, 74…
#>  4 200000628 150110000 ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 83…
#>  5 200000800 338810000 ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 63…
#>  6 200000925 204390000 ((863902.9 6673275, 863187.5 6672486, 861559.8 6671802, …
#>  7 200000933 970790000 ((649872.8 6684591, 649453.9 6685382, 648421.6 6686474, …
#>  8 200004802 401860000 ((1015635 6282628, 1015562 6281731, 1014927 6280935, 101…
#>  9 200005932 414980000 ((608056.9 6735353, 608691.2 6735824, 607865.9 6736734, …
#> 10 200005957 328760000 ((972202.6 6803084, 971065.4 6803547, 970270.7 6803515, …
#> # … with 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                               <GEOMETRY [m]>
#>  1 01    5762580000 POLYGON ((898836.7 6511226, 898866.7 6511939, 898489.7 6512…
#>  2 02    7361650000 POLYGON ((727171.5 6867304, 727564.5 6867911, 726926 686987…
#>  3 03    7340110000 POLYGON ((748815.1 6541775, 748413.8 6543095, 747471.2 6544…
#>  4 04    6925220000 POLYGON ((915178.5 6304594, 914669.9 6304763, 913802.1 6305…
#>  5 05    5527840000 POLYGON ((913795.6 6349572, 914132.6 6351611, 914238.7 6352…
#>  6 06    4298580000 MULTIPOLYGON (((1014207 6277606, 1013506 6278106, 1014114 6…
#>  7 07    5528640000 POLYGON ((788896.9 6359432, 788462.6 6359716, 788033.7 6360…
#>  8 08    5229410000 POLYGON ((799892.1 6914380, 800024.1 6914674, 799092.6 6914…
#>  9 09    4889920000 POLYGON ((596281.9 6168818, 596173.5 6169216, 595071.7 6169…
#> 10 10    6004160000 POLYGON ((781579.9 6760663, 781350.7 6761074, 780472 676083…
#> # … with 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12012270000 POLYGON ((642366.7 6897597, 643111.3 6898042, 644154.7 689…
#>  2 24    39144780000 POLYGON ((543592.6 6610704, 542366.8 6611048, 542056.1 661…
#>  3 27    47784290000 POLYGON ((870726.6 6598915, 869913 6598903, 869203.6 66008…
#>  4 28    29896180000 POLYGON ((472814.1 6821123, 472243.3 6821096, 470192.3 682…
#>  5 32    31795910000 POLYGON ((746069.1 6882373, 747161.7 6882409, 747514.6 688…
#>  6 44    57426620000 POLYGON ((1014643 6714741, 1014629 6715985, 1015421 671816…
#>  7 52    32073040000 MULTIPOLYGON (((376403.9 6582990, 374403.1 6584017, 373253…
#>  8 53    27207910000 MULTIPOLYGON (((286452.7 6721215, 286950.3 6722226, 286095…
#>  9 75    84060310000 MULTIPOLYGON (((406028.7 6195912, 405450.2 6196208, 405482…
#> 10 76    72723600000 MULTIPOLYGON (((636012.3 6147620, 634589 6146717, 632427.7…
#> 11 84    69759110000 POLYGON ((890568.7 6344417, 890721.4 6345901, 890333 63465…
#> 12 93    31378930000 MULTIPOLYGON (((946208.1 6224623, 945800.7 6225179, 946409…
#> 13 94     8679790000 POLYGON ((1223853 6060827, 1223891 6060223, 1223285 605988…
#>