Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34816 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,816 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     <dbl>                                            <MULTIPOLYGON [m]>
#>  1 71010  37130000 (((808611.8 6647932, 808615.2 6646559, 809765.1 6645992, 810…
#>  2 27658   5410000 (((547061.5 6891569, 546785.9 6893178, 546591.2 6894473, 548…
#>  3 08466   7640000 (((855756.2 6944043, 855092.7 6945637, 853332.8 6945791, 853…
#>  4 70017  13910000 (((935224.3 6754156, 935094 6753952, 934785.6 6753695, 93278…
#>  5 01138  29660000 (((917787.2 6532870, 917336.2 6531194, 916959.3 6529439, 916…
#>  6 88142   9160000 (((949816.5 6798005, 950410.1 6798780, 951140.4 6798408, 951…
#>  7 02564  16290000 (((724131 6912396, 721966.2 6912511, 721397.5 6911227, 71997…
#>  8 50634   2510000 (((383001 6948531, 383092.2 6949717, 384020.6 6950756, 38456…
#>  9 07204  48560000 (((812928.8 6432246, 810932.3 6432100, 809501.5 6432794, 807…
#> 10 38052  35750000 (((934010.1 6448769, 934142.3 6451182, 936233.2 6451362, 937…
#> # ℹ 34,806 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111263.9 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150670000 ((966748.4 6552733, 966712.4 6551457, 967694.8 6550219, …
#>  2 200000438 326800000 ((319454 6714465, 321096.5 6713767, 320067.2 6711948, 32…
#>  3 200000545 104800000 ((750344.5 6818238, 749128 6818118, 748496.8 6817507, 74…
#>  4 200000628 150110000 ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 83…
#>  5 200000800 338810000 ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 63…
#>  6 200000925 204390000 ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750, …
#>  7 200000933 970790000 ((649872.8 6684591, 648421.6 6686474, 649133 6688238, 64…
#>  8 200004802 401860000 ((1015562 6281731, 1014114 6280342, 1013506 6278106, 101…
#>  9 200005932 414980000 ((608691.2 6735824, 607263.9 6738882, 606471.7 6738500, …
#> 10 200005957 328760000 ((970270.7 6803515, 968680.1 6804512, 967092.4 6803634, …
#> # ℹ 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5762580000 ((864753.7 6525714, 863508.9 6525325, 863172.2 6525941, 862…
#>  2 02    7361650000 ((731163.3 6862991, 729920.2 6862660, 729324.8 6862996, 729…
#>  3 03    7340110000 ((746447.1 6546598, 745506.5 6547343, 744504.1 6546707, 743…
#>  4 04    6925220000 ((913518.7 6306495, 912835.8 6306600, 911628.6 6306848, 911…
#>  5 05    5527840000 ((914132.6 6351611, 913402.6 6353234, 913478.8 6354443, 914…
#>  6 06    4298580000 ((1014114 6280342, 1015562 6281731, 1015635 6282628, 101483…
#>  7 07    5528640000 ((788033.7 6360072, 787160 6360093, 786566.1 6359438, 78561…
#>  8 08    5229410000 ((799092.6 6914288, 798475.4 6913799, 796762.7 6915021, 795…
#>  9 09    4889920000 ((596281.9 6168818, 596173.5 6169216, 594322.2 6169186, 593…
#> 10 10    6004160000 ((777953.3 6758897, 777007.1 6759702, 776560.5 6759189, 775…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12012270000 POLYGON ((642366.7 6897597, 643111.3 6898042, 644154.7 689…
#>  2 24    39144780000 POLYGON ((542366.8 6611048, 542133.4 6612548, 540477.8 661…
#>  3 27    47784290000 POLYGON ((870726.6 6598915, 869913 6598903, 870030.5 65995…
#>  4 28    29896180000 POLYGON ((473619.7 6821776, 472865.1 6821834, 472243.3 682…
#>  5 32    31795910000 POLYGON ((749282.9 6880399, 749375.5 6879302, 748676.2 687…
#>  6 44    57426620000 POLYGON ((1015421 6718167, 1016315 6718536, 1016147 671884…
#>  7 52    32073040000 MULTIPOLYGON (((371040.1 6587593, 370265.3 6588057, 369077…
#>  8 53    27207910000 MULTIPOLYGON (((267337.9 6730985, 269412.2 6732332, 269637…
#>  9 75    84060310000 MULTIPOLYGON (((406069 6195920, 405384.3 6196171, 405251.9…
#> 10 76    72723600000 MULTIPOLYGON (((636012.3 6147620, 634164.3 6146608, 632427…
#> 11 84    69759110000 POLYGON ((890562.3 6347555, 889905.4 6347996, 888164 63489…
#> 12 93    31378930000 MULTIPOLYGON (((939325.6 6223613, 937269.1 6222913, 936578…
#> 13 94     8679790000 POLYGON ((1222700 6059057, 1221043 6058429, 1221442 605676…
#>