Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34746 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,746 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     [m^2]                                            <MULTIPOLYGON [m]>
#>  1 70177   9130000 (((921738.8 6760207, 921659.6 6760207, 921189.9 6760832, 922…
#>  2 89050   4020000 (((741563.4 6758626, 741320.1 6758008, 739273.9 6757967, 738…
#>  3 60414  10660000 (((660340 6905721, 660211.1 6905210, 659332 6905186, 658695.…
#>  4 80756   3670000 (((661403.5 7000774, 661018.1 7002027, 660347.7 7002280, 660…
#>  5 14478  10140000 (((507839.1 6886562, 509195.2 6884716, 510632.4 6884522, 510…
#>  6 27219   4410000 (((542995.8 6900422, 543112.4 6900476, 543708.9 6900377, 545…
#>  7 68149   2860000 (((1044950 6732060, 1044491 6730419, 1043757 6729491, 104300…
#>  8 30249   3420000 (((797259.9 6301672, 797651.4 6302046, 798774.4 6302539, 799…
#>  9 72363  29160000 (((516821.2 6779360, 515867.3 6781101, 514762.1 6780382, 514…
#> 10 27396   4170000 (((585658.9 6918990, 585542.5 6918351, 584672.7 6918058, 583…
#> # ℹ 34,736 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111282 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150670000 ((966748.4 6552733, 966712.4 6551457, 967694.8 6550219, …
#>  2 200000438 326800000 ((319454 6714465, 321096.5 6713767, 320067.2 6711948, 32…
#>  3 200000545 104800000 ((750344.5 6818238, 749128 6818118, 748496.8 6817507, 74…
#>  4 200000628 150110000 ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 83…
#>  5 200000800 338810000 ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 63…
#>  6 200000925 204390000 ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750, …
#>  7 200000933 970790000 ((649872.8 6684591, 648421.6 6686474, 649133 6688238, 64…
#>  8 200004802 401860000 ((1015562 6281731, 1014114 6280342, 1013506 6278106, 101…
#>  9 200005932 414980000 ((608691.2 6735824, 607263.9 6738882, 606471.7 6738500, …
#> 10 200005957 328760000 ((970270.7 6803515, 968680.1 6804512, 967092.4 6803634, …
#> # ℹ 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5762390000 ((864753.7 6525714, 863508.9 6525325, 863172.2 6525941, 862…
#>  2 02    7361650000 ((731163.3 6862991, 729920.2 6862660, 729324.8 6862996, 729…
#>  3 03    7340110000 ((746447.1 6546598, 745506.5 6547343, 744504.1 6546707, 743…
#>  4 04    6925220000 ((913518.7 6306495, 912835.8 6306600, 911628.6 6306848, 911…
#>  5 05    5548680000 ((914132.6 6351611, 913402.6 6353234, 913478.8 6354443, 914…
#>  6 06    4298580000 ((1016451 6273923, 1015748 6274017, 1014116 6274778, 101490…
#>  7 07    5528640000 ((788033.7 6360072, 787160 6360093, 786566.1 6359438, 78561…
#>  8 08    5228060000 ((799092.6 6914288, 798475.4 6913799, 796762.7 6915021, 795…
#>  9 09    4889920000 ((595335.6 6167604, 596281.9 6168818, 596173.5 6169216, 594…
#> 10 10    6004160000 ((779069.5 6759565, 777953.3 6758897, 777007.1 6759702, 776…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99088.1 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12012270000 POLYGON ((642366.7 6897597, 643111.3 6898042, 644154.7 689…
#>  2 24    39145610000 POLYGON ((542366.8 6611048, 542133.4 6612548, 540477.8 661…
#>  3 27    47783300000 POLYGON ((863680.5 6603221, 863641 6602376, 861623.4 66012…
#>  4 28    29875130000 POLYGON ((473619.7 6821776, 472865.1 6821834, 472243.3 682…
#>  5 32    31815620000 POLYGON ((727076.3 6866042, 727564.5 6867911, 726926 68698…
#>  6 44    57439470000 POLYGON ((1015421 6718167, 1016315 6718536, 1016147 671884…
#>  7 52    32100660000 MULTIPOLYGON (((411461.3 6585132, 410840.4 6586001, 409584…
#>  8 53    27207910000 MULTIPOLYGON (((286722.7 6725074, 287611.2 6724914, 288375…
#>  9 75    84035740000 MULTIPOLYGON (((406069 6195920, 405384.3 6196171, 405482.5…
#> 10 76    72723600000 MULTIPOLYGON (((612901.2 6149983, 612177.2 6151085, 610712…
#> 11 84             NA POLYGON ((890562.3 6347555, 889905.4 6347996, 888164 63489…
#> 12 93    31399770000 MULTIPOLYGON (((935203.8 6224052, 933354.1 6221785, 933074…
#> 13 94     8679790000 MULTIPOLYGON (((1192260 6190036, 1193682 6190244, 1194440 …
#>