Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34746 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,746 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     [m^2]                                                 <POLYGON [m]>
#>  1 70177   9130000 ((921738.8 6760207, 922020.2 6761183, 922906.8 6761192, 9241…
#>  2 89050   4020000 ((741563.4 6758626, 740538.2 6758727, 740595.4 6760220, 7395…
#>  3 60414  10660000 ((660340 6905721, 660705.7 6907293, 659664.3 6907737, 658907…
#>  4 80756   3670000 ((661403.5 7000774, 662247 7001128, 662372.1 7002069, 662035…
#>  5 14478  10140000 ((507839.1 6886562, 508187.4 6885226, 507560.8 6884178, 5073…
#>  6 27219   4410000 ((542995.8 6900422, 543925.7 6899673, 544792.9 6899671, 5463…
#>  7 68149   2860000 ((1044950 6732060, 1044486 6732286, 1043001 6731158, 1043757…
#>  8 30249   3420000 ((797259.9 6301672, 797457.1 6300418, 799608 6300670, 798774…
#>  9 72363  29160000 ((516821.2 6779360, 517104.6 6779775, 518817.2 6781957, 5172…
#> 10 27396   4170000 ((585658.9 6918990, 583762.4 6920427, 583852.4 6919279, 5829…
#> # ℹ 34,736 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111329.5 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150670000 ((967694.8 6550219, 966871.6 6549618, 966610.9 6548763, …
#>  2 200000438 326800000 ((312908.5 6710510, 312603 6709369, 311934.8 6709729, 31…
#>  3 200000545 104800000 ((749128 6818118, 748496.8 6817507, 748719.7 6818137, 74…
#>  4 200000628 150110000 ((841981.7 6348616, 842129.7 6347344, 841340.9 6347339, …
#>  5 200000800 338810000 ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 63…
#>  6 200000925 204390000 ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750, …
#>  7 200000933 970790000 ((658116.6 6688458, 657680.9 6686697, 654603.5 6686858, …
#>  8 200004802 401860000 ((1006225 6288900, 1006943 6288260, 1007710 6288720, 100…
#>  9 200005932 414980000 ((608056.9 6735353, 608691.2 6735824, 607263.9 6738882, …
#> 10 200005957 328760000 ((964840.4 6806088, 963976.2 6807368, 963012.5 6808171, …
#> # ℹ 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5762390000 ((836447.9 6538772, 835167.2 6539004, 834133.8 6539425, 833…
#>  2 02    7361650000 ((712067.1 6888906, 710504.2 6889186, 709997.6 6889771, 709…
#>  3 03    7340110000 ((666522.1 6560462, 666004.6 6563892, 664797.9 6564077, 663…
#>  4 04    6925220000 ((916159.9 6302361, 915178.5 6304594, 913802.1 6305681, 913…
#>  5 05    5548680000 ((916868.6 6348200, 916613.3 6347650, 915174 6346721, 91385…
#>  6 06    4298580000 ((1005824 6290060, 1004519 6291871, 1003358 6292416, 100323…
#>  7 07    5528640000 ((788462.6 6359716, 788033.7 6360072, 787160 6360093, 78656…
#>  8 08    5228060000 ((791218.8 6919801, 790282.6 6920806, 790106.9 6921038, 788…
#>  9 09    4889920000 ((549585.3 6180563, 548401.5 6181779, 546936.7 6182805, 546…
#> 10 10    6004160000 ((755041.6 6781593, 755024.3 6782295, 756156.1 6783595, 755…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99164 ymin: 6049774 xmax: 1242419 ymax: 7110497
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG           AREA                                                   geometry
#>  * <fct>        [m^2]                                             <GEOMETRY [m]>
#>  1 11    12012270000  POLYGON ((603511 6836016, 603140.4 6836375, 601706.7 6835…
#>  2 24    39145610000  POLYGON ((495148.7 6657993, 495492.5 6659013, 495455 6660…
#>  3 27    47783300000  POLYGON ((703731.2 6628693, 702866.6 6630303, 702445.6 66…
#>  4 28    29875130000  POLYGON ((366802.7 6835096, 366610.6 6835047, 366743.4 68…
#>  5 32    31815620000  POLYGON ((607721.5 6901990, 608237.8 6903395, 605920.9 69…
#>  6 44    57439470000  POLYGON ((758919 6780000, 758079.9 6780274, 757175.7 6781…
#>  7 52    32100660000  MULTIPOLYGON (((331680.4 6610035, 331097 6610579, 330522.…
#>  8 53    27207910000  MULTIPOLYGON (((150885.2 6773349, 151308.7 6774432, 15124…
#>  9 75    84035740000  MULTIPOLYGON (((341100.6 6237138, 340672.7 6237834, 34143…
#> 10 76    72723600000  MULTIPOLYGON (((455951.4 6182703, 454293.1 6181044, 45252…
#> 11 84    69789386636. POLYGON ((631741.8 6399699, 631084.9 6400391, 631177.9 64…
#> 12 93    31399770000  MULTIPOLYGON (((848031.6 6250180, 845329 6251305, 843019.…
#> 13 94     8679790000  MULTIPOLYGON (((1182447 6085188, 1181095 6085529, 1181795…
#>