Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34806 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 34,806 × 3
#>    DEPCOM      AREA                                                     geometry
#>  * <chr>      <dbl>                                           <MULTIPOLYGON [m]>
#>  1 74307   2948463. (((941799.9 6548532, 942033.1 6550392, 942537.5 6550435, 94…
#>  2 74098   4706828. (((941651.5 6547465, 942159.6 6546981, 941383.3 6546013, 94…
#>  3 50272  17596289. (((370714.7 6881902, 370172 6881054, 370688.4 6880380, 3695…
#>  4 16206  11304029. (((482978.9 6538251, 483121.9 6537166, 484203.9 6536397, 48…
#>  5 09056  11084393. (((583440.7 6227933, 582467.7 6228255, 582094.6 6229301, 58…
#>  6 32357   6850035. (((538234.5 6295929, 537730.3 6295947, 536490.5 6294652, 53…
#>  7 30004  12057930. (((795990.3 6290300, 796431 6290908, 796081.2 6292583, 7940…
#>  8 25387  12703968. (((989135.6 6693417, 987893.9 6694319, 987417.1 6694984, 98…
#>  9 10116   2113703. (((776224.9 6781924, 775717.7 6782077, 775132.8 6783433, 77…
#> 10 13109  10854229. (((900537.1 6271310, 900284.8 6272182, 902540.3 6273213, 90…
#> # ℹ 34,796 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111263.9 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI            AREA                                                 geometry
#>    <fct>          [m^2]                                            <POLYGON [m]>
#>  1 200000172 150930833. ((959255.3 6548853, 958947.9 6550658, 958574 6552122, 9…
#>  2 200000438 327150254. ((319454 6714465, 321096.5 6713767, 320067.2 6711948, 3…
#>  3 200000545 104902649. ((750344.5 6818238, 749128 6818118, 748496.8 6817507, 7…
#>  4 200000628 150545742. ((841106 6346067, 839847.8 6345984, 839321.7 6344038, 8…
#>  5 200000800 342496789. ((631021.1 6720275, 629746 6718932, 631310.3 6717273, 6…
#>  6 200000925 204174303. ((861559.8 6671802, 860515.3 6673049, 859548.4 6672750,…
#>  7 200000933 974245358. ((649872.8 6684591, 648421.6 6686474, 649133 6688238, 6…
#>  8 200004802 408212483. ((1015562 6281731, 1014114 6280342, 1013506 6278106, 10…
#>  9 200005932 417348141. ((608691.2 6735824, 607263.9 6738882, 606471.7 6738500,…
#> 10 200005957 328868906. ((970270.7 6803515, 968680.1 6804512, 967092.4 6803634,…
#> # ℹ 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP          AREA                                                    geometry
#>  * <fct>       [m^2]                                               <POLYGON [m]>
#>  1 01    5774236437. ((864753.7 6525714, 863508.9 6525325, 863172.2 6525941, 86…
#>  2 02    7418547459. ((731163.3 6862991, 729920.2 6862660, 729324.8 6862996, 72…
#>  3 03    7365661052. ((746447.1 6546598, 745506.5 6547343, 744504.1 6546707, 74…
#>  4 04    6994180808. ((913518.7 6306495, 912835.8 6306600, 911628.6 6306848, 91…
#>  5 05    5685034441. ((914132.6 6351611, 913402.6 6353234, 913478.8 6354443, 91…
#>  6 06    4294743462. ((1014114 6280342, 1015562 6281731, 1015635 6282628, 10148…
#>  7 07    5561382596. ((788033.7 6360072, 787160 6360093, 786566.1 6359438, 7856…
#>  8 08    5251958197. ((799092.6 6914288, 798475.4 6913799, 796762.7 6915021, 79…
#>  9 09    4921161085. ((596281.9 6168818, 596173.5 6169216, 594322.2 6169186, 59…
#> 10 10    6022295019. ((779069.5 6759565, 777953.3 6758897, 777007.1 6759702, 77…
#> # ℹ 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049731 xmax: 1242440 ymax: 7110479
#> Projected CRS: RGF93 v1 / Lambert-93
#> # A tibble: 13 × 3
#>    REG           AREA                                                   geometry
#>  * <fct>        [m^2]                                             <GEOMETRY [m]>
#>  1 11    12064959812. POLYGON ((642366.7 6897597, 643111.3 6898042, 644154.7 68…
#>  2 24    39470063209. POLYGON ((542366.8 6611048, 542133.4 6612548, 540477.8 66…
#>  3 27    47980510796. POLYGON ((870726.6 6598915, 869913 6598903, 870030.5 6599…
#>  4 28    30099799337. POLYGON ((473619.7 6821776, 472865.1 6821834, 472243.3 68…
#>  5 32    32010878443. POLYGON ((749282.9 6880399, 749375.5 6879302, 748676.2 68…
#>  6 44    57704251724. POLYGON ((1015421 6718167, 1016315 6718536, 1016147 67188…
#>  7 52    32359129250. MULTIPOLYGON (((403893.1 6586987, 402889.7 6587182, 40107…
#>  8 53    27445932632. MULTIPOLYGON (((269637.9 6734259, 270864.3 6735502, 27050…
#>  9 75    85093927877. MULTIPOLYGON (((406069 6195920, 405384.3 6196171, 405251.…
#> 10 76    73405626578. MULTIPOLYGON (((636012.3 6147620, 634164.3 6146608, 63242…
#> 11 84    70796980594. POLYGON ((890562.3 6347555, 889905.4 6347996, 888164 6348…
#> 12 93    31674502874. MULTIPOLYGON (((925820.4 6229158, 926292.6 6230024, 92547…
#> 13 94     8754670769. POLYGON ((1222700 6059057, 1221043 6058429, 1221442 60567…
#>