Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom

la commune sur laquelle filtrer les données

epci

l'epci sur lequel filtrer les données

dep

le departement sur lequel filtrer les données

reg

la region sur laquelle filtrer les données

Value

une liste de spatial dataframe (sf)

Examples

get_map(epci = "244400404")
#> $com_geo
#> Simple feature collection with 34826 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 34,826 × 3
#>    DEPCOM     AREA                                                      geometry
#>  * <chr>     <dbl>                                            <MULTIPOLYGON [m]>
#>  1 43076  21890000 (((748166.1 6463827, 747512.2 6465082, 748162.4 6465513, 748…
#>  2 65462  16010000 (((455022.6 6263682, 454435.4 6264116, 453913.1 6262475, 453…
#>  3 29058  32760000 (((178183.7 6781582, 177738.5 6780262, 178285 6780064, 17834…
#>  4 22218  15540000 (((243953.6 6877260, 243491.9 6876618, 243054.9 6876860, 241…
#>  5 38250   4900000 (((889525.8 6504615, 889698.2 6504084, 888708.2 6503830, 888…
#>  6 86097  27220000 (((496845 6582010, 497689.1 6583549, 498147.7 6583726, 49855…
#>  7 27058  36530000 (((574623 6898578, 573804.1 6899465, 573913.6 6900680, 57364…
#>  8 42225   8670000 (((821353.7 6493133, 822119.5 6492921, 822431 6493626, 82175…
#>  9 65478   6520000 (((451686.2 6205143, 450702.1 6205205, 449116.3 6206205, 449…
#> 10 06043  17140000 (((1046530 6320331, 1046759 6319283, 1047343 6318288, 104736…
#> # … with 34,816 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111318.1 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 1,232 × 3
#>    EPCI           AREA                                                  geometry
#>    <fct>         [m^2]                                             <POLYGON [m]>
#>  1 200000172 150670000 ((967057.6 6556353, 967237.5 6555934, 966748.4 6552733, …
#>  2 200000438 326800000 ((325038.9 6719791, 325279 6718117, 325297.8 6716916, 32…
#>  3 200000545 104800000 ((751948.9 6819322, 750344.5 6818238, 749128 6818118, 74…
#>  4 200000628 150110000 ((842129.7 6347344, 841340.9 6347339, 841106 6346067, 83…
#>  5 200000800 338810000 ((631434.7 6721265, 631021.1 6720275, 629746 6718932, 63…
#>  6 200000925 204390000 ((864862.5 6674862, 864543.7 6674489, 864723.7 6673592, …
#>  7 200000933 970790000 ((652434.2 6683159, 652145.3 6683270, 651352.4 6682885, …
#>  8 200004802 401860000 ((994912.4 6275380, 994837.6 6276438, 994254.9 6277375, …
#>  9 200005932 414980000 ((603944.5 6733623, 604602.7 6734013, 605795.4 6734566, …
#> 10 200005957 328760000 ((974704.6 6803983, 973836.4 6803884, 972968.9 6803285, …
#> # … with 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP         AREA                                                     geometry
#>  * <fct>      [m^2]                                               <GEOMETRY [m]>
#>  1 01    5762580000 POLYGON ((904448.6 6504924, 903715.4 6506249, 903501 650678…
#>  2 02    7361650000 POLYGON ((742104 6871179, 741982.8 6870011, 742238.7 686940…
#>  3 03    7340110000 POLYGON ((753778.8 6537018, 752894.5 6538126, 752435.3 6539…
#>  4 04    6925220000 POLYGON ((943701.1 6290140, 942852.7 6290779, 942340.2 6292…
#>  5 05    5527840000 POLYGON ((920082 6349463, 919321 6348758, 918436.6 6348360,…
#>  6 06    4298580000 MULTIPOLYGON (((1018233 6272508, 1016644 6272701, 1016451 6…
#>  7 07    5528640000 POLYGON ((802747.3 6358147, 802960 6357489, 802811.9 635558…
#>  8 08    5229410000 POLYGON ((823029.8 6907777, 822289.8 6907219, 821646 690590…
#>  9 09    4889920000 POLYGON ((603392.4 6165331, 601015.4 6164281, 600207.2 6164…
#> 10 10    6004160000 POLYGON ((796986.4 6761731, 797284.4 6760524, 796546.8 6759…
#> # … with 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99041.4 ymin: 6049643 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 13 × 3
#>    REG          AREA                                                    geometry
#>  * <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12012270000 POLYGON ((660728.7 6783927, 661214.5 6784158, 661574 67850…
#>  2 24    39144780000 POLYGON ((633985.7 6591608, 632939.8 6591403, 631793.8 659…
#>  3 27    47784290000 POLYGON ((787479.5 6564794, 787419.1 6564795, 787418 65651…
#>  4 28    29896180000 POLYGON ((530883.8 6791099, 530747.4 6791627, 530350.2 679…
#>  5 32    31795910000 POLYGON ((742104 6871179, 741982.8 6870011, 742238.7 68694…
#>  6 44    57426620000 POLYGON ((1033584 6715931, 1033526 6715303, 1032958 671434…
#>  7 52    32073040000 MULTIPOLYGON (((414138 6586391, 413264 6584951, 411461.3 6…
#>  8 53    27207910000 MULTIPOLYGON (((291878.6 6722561, 291032.4 6722104, 290708…
#>  9 75    84060310000 MULTIPOLYGON (((413496.1 6198389, 412489.3 6197827, 411470…
#> 10 76    72723600000 MULTIPOLYGON (((671874.6 6142280, 672181.2 6141424, 672020…
#> 11 84    69759110000 POLYGON ((896441.6 6338627, 896106.3 6338849, 895821.8 634…
#> 12 93    31378930000 MULTIPOLYGON (((988888.1 6238512, 988006.5 6237917, 987440…
#> 13 94     8679790000 POLYGON ((1226006 6063425, 1225792 6062710, 1224632 606205…
#>