Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom
la commune sur laquelle filtrer les données

epci
l'epci sur lequel filtrer les données

dep
le departement sur lequel filtrer les données

reg
la region sur laquelle filtrer les données

Value

une liste de spatial dataframe

Examples

get_map(epci = "244400404")

#> Warning: argument is not an atomic vector; coercing

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> $com_geo
#> Simple feature collection with 34839 features and 2 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> projected CRS:  RGF93 / Lambert-93
#> # A tibble: 34,839 x 3
#>    DEPCOM     AREA                                                      geometry
#>    <fct>     [m^2]                                            <MULTIPOLYGON [m]>
#>  1 32216   9376180 (((501878.2 6267455, 501951 6267382, 502074.3 6266956, 50242…
#>  2 47033  10254007 (((517154.1 6383591, 516723 6383480, 516671.1 6383457, 51660…
#>  3 38225  77705208 (((902147.8 6463688, 902165.6 6463719, 902290 6463767, 90234…
#>  4 32009   9185168 (((473133.5 6280663, 473157.5 6280674, 473771.2 6280615, 474…
#>  5 62890  10276478 (((641994.8 7030576, 641859.2 7030505, 641751.8 7030348, 641…
#>  6 08018   4399794 (((823029.8 6907777, 823097.1 6907892, 823421.8 6908346, 823…
#>  7 32113   9276726 (((461213.2 6299858, 461182.4 6299868, 460553.2 6300193, 460…
#>  8 10319  19210557 (((744144.3 6786646, 744012.9 6786778, 743775.6 6786895, 743…
#>  9 42070  26714264 (((779266.8 6538044, 779306.7 6538065, 779341 6538063, 77946…
#> 10 06097  22202325 (((1025406 6315941, 1025281 6316407, 1025230 6316956, 102532…
#> # … with 34,829 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1232 features and 2 fields
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 111367.9 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> projected CRS:  RGF93 / Lambert-93
#> # A tibble: 1,232 x 3
#>    EPCI          AREA                                                   geometry
#>    <fct>        [m^2]                                              <POLYGON [m]>
#>  1 2000001… 150696248 ((966748.4 6552733, 966755.8 6552381, 966730.4 6551740, 9…
#>  2 2000004… 327033526 ((310698.9 6711370, 309865.1 6712477, 309769.4 6712376, 3…
#>  3 2000005… 105210176 ((748496.8 6817507, 748719.7 6818137, 748608.2 6818254, 7…
#>  4 2000006… 150413311 ((836207.1 6345747, 836195.7 6345851, 836197.2 6345973, 8…
#>  5 2000008… 342468283 ((631434.7 6721265, 631378 6721142, 631296.9 6720920, 631…
#>  6 2000009… 204165930 ((864723.7 6673592, 864701.5 6673591, 864660 6673555, 864…
#>  7 2000009… 868359249 ((657974.2 6689616, 658009.5 6689367, 658084.8 6689146, 6…
#>  8 2000048… 408046364 ((1005237 6278923, 1005251 6278705, 1005139 6278186, 1005…
#>  9 2000059… 417855138 ((604602.7 6734013, 604749 6734098, 605124.3 6734242, 605…
#> 10 2000059… 328763775 ((980306.8 6808608, 980228.7 6808502, 980198.4 6808427, 9…
#> # … with 1,222 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> projected CRS:  RGF93 / Lambert-93
#> # A tibble: 96 x 3
#>    DEP         AREA                                                     geometry
#>    <fct>      [m^2]                                                <POLYGON [m]>
#>  1 01    5773773857 ((904448.6 6504924, 904395.9 6505018, 904252.4 6505282, 904…
#>  2 02    7418972784 ((733174.4 6861885, 733166.1 6861912, 732924.7 6862170, 732…
#>  3 03    7365259270 ((754951.6 6541859, 754836.1 6541737, 754334 6541292, 75434…
#>  4 04    6993793070 ((946374.1 6293789, 946188.4 6293686, 946082.4 6293625, 946…
#>  5 05    5685308934 ((923049.9 6349816, 922918.4 6349761, 922468 6349539, 92243…
#>  6 06    4291617476 ((1018318 6276621, 1018325 6276578, 1018381 6275907, 101838…
#>  7 07    5562045799 ((802811.9 6355587, 802576.9 6355227, 802267.4 6354729, 802…
#>  8 08    5253130973 ((805851.6 6909444, 805657 6909338, 805527.8 6909247, 80533…
#>  9 09    4921753472 ((576353.1 6173801, 576341.6 6173825, 576130.4 6173783, 575…
#> 10 10    6021833414 ((796986.4 6761731, 797037.4 6761658, 797073 6761560, 79708…
#> # … with 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> projected CRS:  RGF93 / Lambert-93
#> # A tibble: 13 x 3
#>    REG          AREA                                                    geometry
#>    <fct>       [m^2]                                              <GEOMETRY [m]>
#>  1 11    12065541081 POLYGON ((664448.3 6780657, 664374.6 6780628, 664258.8 678…
#>  2 24    39471195196 POLYGON ((636117.2 6591894, 635918.3 6591879, 635321.7 659…
#>  3 27    47980633474 POLYGON ((787479.5 6564794, 787419.1 6564795, 787365.2 656…
#>  4 28    30100397503 POLYGON ((535826.7 6798024, 535824.2 6797962, 535837.1 679…
#>  5 32    32008275180 POLYGON ((733174.4 6861885, 733166.1 6861912, 732924.7 686…
#>  6 44    57704056624 POLYGON ((1032761 6718208, 1032852 6718263, 1033261 671842…
#>  7 52    32357466715 MULTIPOLYGON (((411461.3 6585132, 411394.4 6585198, 411181…
#>  8 53    27414290032 MULTIPOLYGON (((238065.3 6709325, 238074.6 6709320, 238747…
#>  9 75    84664419969 MULTIPOLYGON (((413496.1 6198389, 413451.2 6198381, 413114…
#> 10 76    73404792715 MULTIPOLYGON (((671874.6 6142280, 671920.4 6142190, 672065…
#> 11 84    70796442425 POLYGON ((904059.3 6342273, 904060 6342232, 903977.2 63418…
#> 12 93    31633318728 MULTIPOLYGON (((936608.4 6224346, 936651.4 6224328, 936792…
#> 13 94     8748236923 POLYGON ((1226006 6063425, 1225991 6063418, 1225800 606288…
#>