Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM

get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)

Arguments

depcom
la commune sur laquelle filtrer les données

epci
l'epci sur lequel filtrer les données

dep
le departement sur lequel filtrer les données

reg
la region sur laquelle filtrer les données

Value

une liste de spatial dataframe

Examples

get_map(epci = "244400404")

#> Warning: argument is not an atomic vector; coercing

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> Warning: argument is not an atomic vector; coercing

#> $com_geo
#> Simple feature collection with 34836 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 34,836 × 3
#>    DEPCOM      AREA                                                     geometry
#>    <fct>      [m^2]                                           <MULTIPOLYGON [m]>
#>  1 01001  15737871. (((846985.3 6564105, 846995.7 6564208, 847015.2 6564655, 84…
#>  2 01002   9112278. (((886026.1 6548175, 886157.9 6548496, 886285.3 6548697, 88…
#>  3 01004  24784212. (((884889.8 6539631, 884822.4 6539610, 884795.7 6539615, 88…
#>  4 01005  16004376. (((845947 6547998, 846057.5 6548051, 846074.5 6548052, 8466…
#>  5 01006   6092798. (((901335.5 6521127, 901401.1 6521152, 901549.8 6521169, 90…
#>  6 01007  33717139. (((886026.1 6548175, 886141.6 6548106, 886188.1 6548070, 88…
#>  7 01008   5081379. (((878932.6 6540739, 879047.1 6540688, 879288.2 6540766, 87…
#>  8 01009   6930424. (((905769.6 6521658, 905734.7 6521678, 905713.8 6521711, 90…
#>  9 01010  29617065. (((913981.3 6535875, 914037.1 6536215, 914046.2 6536393, 91…
#> 10 01011  15140592. (((903239.9 6568380, 903245 6568389, 903322.7 6568716, 9032…
#> # … with 34,826 more rows
#> 
#> $epci_geo
#> Simple feature collection with 1231 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 111367.9 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 1,231 × 3
#>    EPCI            AREA                                                 geometry
#>    <fct>          [m^2]                                            <POLYGON [m]>
#>  1 200000172 150696248. ((966748.4 6552733, 966755.8 6552381, 966730.4 6551740,…
#>  2 200000438 327033526. ((310698.9 6711370, 309865.1 6712477, 309769.4 6712376,…
#>  3 200000545 105210176. ((748496.8 6817507, 748719.7 6818137, 748608.2 6818254,…
#>  4 200000628 150413311. ((836207.1 6345747, 836195.7 6345851, 836197.2 6345973,…
#>  5 200000800 342468283. ((631434.7 6721265, 631378 6721142, 631296.9 6720920, 6…
#>  6 200000925 204165930. ((864723.7 6673592, 864701.5 6673591, 864660 6673555, 8…
#>  7 200000933 974277003. ((657974.2 6689616, 658009.5 6689367, 658084.8 6689146,…
#>  8 200004802 408046364. ((1005237 6278923, 1005251 6278705, 1005139 6278186, 10…
#>  9 200005932 417855138. ((604602.7 6734013, 604749 6734098, 605124.3 6734242, 6…
#> 10 200005957 328763775. ((980306.8 6808608, 980228.7 6808502, 980198.4 6808427,…
#> # … with 1,221 more rows
#> 
#> $dep_geo
#> Simple feature collection with 96 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 96 × 3
#>    DEP          AREA                                                    geometry
#>    <fct>       [m^2]                                               <POLYGON [m]>
#>  1 01    5773773857. ((904448.6 6504924, 904395.9 6505018, 904252.4 6505282, 90…
#>  2 02    7418972784. ((733174.4 6861885, 733166.1 6861912, 732924.7 6862170, 73…
#>  3 03    7365259270. ((754951.6 6541859, 754836.1 6541737, 754334 6541292, 7543…
#>  4 04    6993793070. ((946374.1 6293789, 946188.4 6293686, 946082.4 6293625, 94…
#>  5 05    5685308934. ((923049.9 6349816, 922918.4 6349761, 922468 6349539, 9224…
#>  6 06    4291617476. ((1018318 6276621, 1018325 6276578, 1018381 6275907, 10183…
#>  7 07    5562045799. ((802811.9 6355587, 802576.9 6355227, 802267.4 6354729, 80…
#>  8 08    5253130973. ((805851.6 6909444, 805657 6909338, 805527.8 6909247, 8053…
#>  9 09    4921753472. ((576353.1 6173801, 576341.6 6173825, 576130.4 6173783, 57…
#> 10 10    6021833414. ((796986.4 6761731, 797037.4 6761658, 797073 6761560, 7970…
#> # … with 86 more rows
#> 
#> $reg_geo
#> Simple feature collection with 13 features and 2 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 99040 ymin: 6049681 xmax: 1242444 ymax: 7110479
#> Projected CRS: RGF93 / Lambert-93
#> # A tibble: 13 × 3
#>    REG           AREA                                                   geometry
#>    <fct>        [m^2]                                             <GEOMETRY [m]>
#>  1 11    12065541081. POLYGON ((664448.3 6780657, 664374.6 6780628, 664258.8 67…
#>  2 24    39471195196. POLYGON ((636117.2 6591894, 635918.3 6591879, 635321.7 65…
#>  3 27    47980633474. POLYGON ((787479.5 6564794, 787419.1 6564795, 787365.2 65…
#>  4 28    30100397503. POLYGON ((535826.7 6798024, 535824.2 6797962, 535837.1 67…
#>  5 32    32008275180. POLYGON ((733174.4 6861885, 733166.1 6861912, 732924.7 68…
#>  6 44    57704056624. POLYGON ((1032761 6718208, 1032852 6718263, 1033261 67184…
#>  7 52    32357466715. MULTIPOLYGON (((411461.3 6585132, 411394.4 6585198, 41118…
#>  8 53    27414290032. MULTIPOLYGON (((238065.3 6709325, 238074.6 6709320, 23874…
#>  9 75    84664419969. MULTIPOLYGON (((413496.1 6198389, 413451.2 6198381, 41311…
#> 10 76    73404792715. MULTIPOLYGON (((671874.6 6142280, 671920.4 6142190, 67206…
#> 11 84    70796442425. POLYGON ((904059.3 6342273, 904060 6342232, 903977.2 6341…
#> 12 93    31633318728. MULTIPOLYGON (((936608.4 6224346, 936651.4 6224328, 93679…
#> 13 94     8748236923. POLYGON ((1226006 6063425, 1225991 6063418, 1225800 60628…
#>