Renvoie la liste des territoires pour un ou des territoires

Renvoie la liste des territoires pour un ou des territoires

Renvoie la liste des départements pour une ou des régions

Renvoie la liste des EPCI pour un département

Renvoie la liste des EPCI pour une région

Renvoie la liste des communes pour un epci

Renvoie la liste des communes pour un ou des départements

Renvoie la liste des communes pour une région

list_dep_in_reg(reg)

list_epci_in_dep(dep)

list_epci_in_reg(reg)

list_com_in_epci(epci)

list_com_in_dep(dep)

list_com_in_reg(reg)

Arguments

reg

Vecteur de caractère. Le code de la région concernée.

dep

Vecteur de caractère. Le code du département concernée.

epci

Vecteur de caractère. Le code de l'epci concernée.

Value

Un vecteur de caractère.

Examples

list_dep_in_reg("75")
#>  [1] "16" "17" "19" "23" "24" "33" "40" "47" "64" "79" "86" "87"
list_epci_in_dep("24")
#>  [1] "200070647" "200040392" "243301454" "200041440" "200041168" "200034197"
#>  [7] "200040889" "200034833" "200041572" "200040830" "242400935" "243301371"
#> [13] "200071819" "200040400" "242401024" "200040384" "200069094" "200040095"
#> [19] "242400752" "200027217" "200041150" "200041051"
list_epci_in_reg("52")
#>  [1] "244400404" "244900015" "246100663" "247200132" "200071678" "200067635"
#>  [7] "244400610" "244400644" "200023778" "248500589" "200083392" "200071165"
#> [13] "200060010" "200067346" "243500741" "200071876" "200070233" "244900809"
#> [19] "200068955" "244900882" "200071629" "200072726" "247200090" "244400503"
#> [25] "245300355" "248500191" "200072718" "244400453" "244400537" "200072882"
#> [31] "200033298" "200072692" "200071868" "245300389" "200042182" "244400552"
#> [37] "245300447" "200048551" "200071934" "247200686" "248500415" "245300223"
#> [43] "248500662" "200000438" "248500464" "200071918" "248500530" "248500621"
#> [49] "247200348" "247200421" "244400586" "247200629" "200072734" "244400438"
#> [55] "200072700" "200072684" "200070373" "200071553" "200040475" "200068963"
#> [61] "200072676" "200055887" "248500258" "247200447" "248500340" "200067866"
#> [67] "200071546" "200073112" "200073260" "200071900" "248500563"
list_com_in_epci("200070647")
#>  [1] "24037" "24051" "24054" "24126" "24140" "24145" "24148" "24182" "24191"
#> [10] "24193" "24194" "24197" "24222" "24224" "24225" "24237" "24246" "24267"
#> [19] "24274" "24276" "24277" "24296" "24331" "24340" "24345" "24349" "24351"
#> [28] "24357" "24413" "24419" "24420" "24437" "24472" "24487" "24499" "24523"
#> [37] "24534" "24549"
list_com_in_dep("24")
#>   [1] "24001" "24002" "24004" "24005" "24006" "24007" "24008" "24009" "24010"
#>  [10] "24011" "24012" "24014" "24015" "24016" "24018" "24019" "24020" "24021"
#>  [19] "24022" "24023" "24024" "24025" "24026" "24027" "24028" "24029" "24030"
#>  [28] "24031" "24032" "24034" "24035" "24036" "24037" "24038" "24039" "24040"
#>  [37] "24042" "24043" "24045" "24046" "24048" "24050" "24051" "24052" "24053"
#>  [46] "24054" "24055" "24056" "24057" "24058" "24059" "24060" "24061" "24062"
#>  [55] "24063" "24064" "24066" "24067" "24068" "24069" "24070" "24071" "24073"
#>  [64] "24074" "24075" "24076" "24077" "24080" "24081" "24082" "24083" "24084"
#>  [73] "24085" "24086" "24087" "24088" "24090" "24091" "24094" "24095" "24096"
#>  [82] "24097" "24098" "24100" "24101" "24102" "24104" "24105" "24106" "24107"
#>  [91] "24108" "24109" "24110" "24111" "24113" "24114" "24115" "24116" "24117"
#> [100] "24119" "24120" "24121" "24122" "24123" "24124" "24126" "24128" "24129"
#> [109] "24130" "24131" "24132" "24133" "24134" "24135" "24136" "24137" "24138"
#> [118] "24139" "24140" "24141" "24142" "24143" "24144" "24145" "24146" "24147"
#> [127] "24148" "24150" "24151" "24152" "24153" "24154" "24155" "24156" "24157"
#> [136] "24158" "24159" "24160" "24161" "24162" "24163" "24164" "24165" "24167"
#> [145] "24168" "24171" "24172" "24174" "24175" "24176" "24177" "24179" "24180"
#> [154] "24182" "24183" "24184" "24186" "24188" "24189" "24190" "24191" "24192"
#> [163] "24193" "24194" "24195" "24196" "24197" "24199" "24200" "24202" "24205"
#> [172] "24206" "24207" "24208" "24209" "24210" "24211" "24212" "24213" "24214"
#> [181] "24215" "24216" "24217" "24218" "24220" "24221" "24222" "24223" "24224"
#> [190] "24225" "24226" "24227" "24228" "24229" "24230" "24231" "24232" "24234"
#> [199] "24236" "24237" "24238" "24240" "24241" "24242" "24243" "24244" "24245"
#> [208] "24246" "24247" "24248" "24251" "24252" "24253" "24254" "24255" "24256"
#> [217] "24257" "24259" "24260" "24261" "24262" "24263" "24264" "24266" "24267"
#> [226] "24268" "24269" "24271" "24272" "24273" "24274" "24276" "24277" "24278"
#> [235] "24279" "24280" "24281" "24282" "24284" "24285" "24286" "24287" "24288"
#> [244] "24289" "24290" "24291" "24292" "24293" "24294" "24295" "24296" "24297"
#> [253] "24299" "24300" "24301" "24302" "24303" "24304" "24305" "24306" "24307"
#> [262] "24308" "24309" "24311" "24312" "24313" "24316" "24317" "24318" "24319"
#> [271] "24320" "24321" "24322" "24323" "24324" "24325" "24326" "24327" "24328"
#> [280] "24329" "24330" "24331" "24334" "24335" "24336" "24337" "24338" "24339"
#> [289] "24340" "24341" "24345" "24346" "24347" "24348" "24349" "24350" "24351"
#> [298] "24352" "24353" "24354" "24355" "24356" "24357" "24359" "24360" "24361"
#> [307] "24362" "24364" "24365" "24366" "24367" "24370" "24371" "24372" "24373"
#> [316] "24374" "24375" "24376" "24377" "24378" "24379" "24380" "24381" "24382"
#> [325] "24383" "24384" "24385" "24386" "24388" "24390" "24392" "24393" "24394"
#> [334] "24395" "24396" "24397" "24398" "24399" "24401" "24403" "24404" "24405"
#> [343] "24406" "24407" "24408" "24409" "24410" "24411" "24412" "24413" "24414"
#> [352] "24415" "24416" "24417" "24418" "24419" "24420" "24421" "24422" "24423"
#> [361] "24424" "24425" "24426" "24428" "24429" "24432" "24434" "24436" "24437"
#> [370] "24438" "24441" "24442" "24443" "24444" "24445" "24446" "24448" "24449"
#> [379] "24450" "24451" "24452" "24453" "24454" "24455" "24456" "24457" "24458"
#> [388] "24459" "24460" "24461" "24462" "24463" "24464" "24465" "24466" "24468"
#> [397] "24470" "24471" "24472" "24473" "24474" "24476" "24477" "24478" "24479"
#> [406] "24480" "24481" "24482" "24483" "24484" "24485" "24486" "24487" "24488"
#> [415] "24489" "24490" "24491" "24492" "24493" "24494" "24495" "24496" "24498"
#> [424] "24499" "24500" "24501" "24502" "24504" "24505" "24507" "24508" "24509"
#> [433] "24510" "24511" "24512" "24513" "24514" "24515" "24516" "24517" "24518"
#> [442] "24519" "24520" "24521" "24522" "24523" "24524" "24525" "24526" "24527"
#> [451] "24528" "24529" "24531" "24532" "24533" "24534" "24535" "24536" "24537"
#> [460] "24538" "24540" "24541" "24542" "24543" "24544" "24545" "24546" "24547"
#> [469] "24548" "24549" "24550" "24551" "24552" "24553" "24554" "24555" "24557"
#> [478] "24558" "24559" "24560" "24562" "24563" "24564" "24565" "24566" "24567"
#> [487] "24568" "24569" "24570" "24571" "24572" "24573" "24574" "24575" "24576"
#> [496] "24577" "24580" "24581" "24582" "24584" "24585" "24586" "24587"
list_com_in_reg("52")
#>    [1] "44001" "44002" "44003" "44005" "44006" "44007" "44009" "44010" "44012"
#>   [10] "44013" "44014" "44015" "44016" "44018" "44019" "44020" "44021" "44022"
#>   [19] "44023" "44024" "44025" "44026" "44027" "44028" "44029" "44030" "44031"
#>   [28] "44032" "44033" "44035" "44036" "44037" "44038" "44039" "44041" "44043"
#>   [37] "44044" "44045" "44046" "44047" "44048" "44049" "44050" "44051" "44052"
#>   [46] "44053" "44054" "44055" "44056" "44057" "44058" "44061" "44062" "44063"
#>   [55] "44064" "44065" "44066" "44067" "44068" "44069" "44070" "44071" "44072"
#>   [64] "44073" "44074" "44075" "44076" "44077" "44078" "44079" "44080" "44081"
#>   [73] "44082" "44083" "44084" "44085" "44086" "44087" "44088" "44089" "44090"
#>   [82] "44091" "44092" "44094" "44095" "44096" "44097" "44098" "44099" "44100"
#>   [91] "44101" "44102" "44103" "44104" "44105" "44106" "44107" "44108" "44109"
#>  [100] "44110" "44111" "44112" "44113" "44114" "44115" "44116" "44117" "44118"
#>  [109] "44119" "44120" "44121" "44122" "44123" "44124" "44125" "44126" "44127"
#>  [118] "44128" "44129" "44130" "44131" "44132" "44133" "44134" "44135" "44136"
#>  [127] "44137" "44138" "44139" "44140" "44141" "44142" "44143" "44144" "44145"
#>  [136] "44146" "44148" "44149" "44150" "44151" "44152" "44153" "44154" "44155"
#>  [145] "44156" "44157" "44158" "44159" "44161" "44162" "44163" "44164" "44165"
#>  [154] "44166" "44168" "44169" "44170" "44171" "44172" "44173" "44174" "44175"
#>  [163] "44176" "44178" "44179" "44180" "44182" "44183" "44184" "44185" "44186"
#>  [172] "44187" "44188" "44189" "44190" "44192" "44193" "44194" "44195" "44196"
#>  [181] "44197" "44198" "44199" "44200" "44201" "44202" "44203" "44204" "44205"
#>  [190] "44206" "44207" "44208" "44209" "44210" "44211" "44212" "44213" "44214"
#>  [199] "44215" "44216" "44217" "44218" "44220" "44221" "44222" "44223" "44224"
#>  [208] "49002" "49003" "49007" "49008" "49009" "49010" "49011" "49012" "49015"
#>  [217] "49017" "49018" "49020" "49021" "49022" "49023" "49026" "49027" "49028"
#>  [226] "49029" "49030" "49035" "49036" "49038" "49041" "49045" "49048" "49050"
#>  [235] "49053" "49054" "49055" "49056" "49057" "49058" "49060" "49061" "49063"
#>  [244] "49064" "49067" "49068" "49069" "49070" "49076" "49080" "49082" "49086"
#>  [253] "49089" "49090" "49092" "49099" "49100" "49102" "49107" "49109" "49110"
#>  [262] "49112" "49113" "49114" "49120" "49121" "49123" "49125" "49127" "49129"
#>  [271] "49130" "49131" "49132" "49135" "49138" "49140" "49155" "49160" "49161"
#>  [280] "49163" "49167" "49170" "49171" "49174" "49176" "49178" "49180" "49182"
#>  [289] "49183" "49188" "49192" "49193" "49194" "49195" "49200" "49201" "49205"
#>  [298] "49209" "49211" "49214" "49215" "49216" "49217" "49218" "49219" "49220"
#>  [307] "49221" "49222" "49223" "49224" "49228" "49231" "49235" "49236" "49237"
#>  [316] "49240" "49241" "49244" "49246" "49247" "49248" "49253" "49257" "49259"
#>  [325] "49260" "49261" "49262" "49266" "49267" "49269" "49271" "49272" "49278"
#>  [334] "49283" "49284" "49288" "49291" "49292" "49294" "49298" "49299" "49301"
#>  [343] "49302" "49306" "49307" "49308" "49310" "49311" "49321" "49323" "49326"
#>  [352] "49328" "49329" "49330" "49331" "49332" "49333" "49334" "49336" "49338"
#>  [361] "49339" "49341" "49343" "49344" "49345" "49347" "49352" "49353" "49355"
#>  [370] "49358" "49359" "49361" "49362" "49364" "49367" "49368" "49369" "49370"
#>  [379] "49371" "49373" "49374" "49377" "49378" "49381" "53001" "53002" "53003"
#>  [388] "53005" "53007" "53008" "53009" "53010" "53011" "53012" "53013" "53015"
#>  [397] "53016" "53017" "53018" "53019" "53021" "53022" "53023" "53025" "53026"
#>  [406] "53027" "53028" "53029" "53030" "53031" "53033" "53034" "53035" "53036"
#>  [415] "53037" "53038" "53039" "53040" "53041" "53042" "53043" "53045" "53046"
#>  [424] "53047" "53048" "53049" "53051" "53052" "53053" "53054" "53055" "53056"
#>  [433] "53057" "53058" "53059" "53061" "53062" "53063" "53064" "53066" "53067"
#>  [442] "53068" "53069" "53071" "53072" "53073" "53074" "53075" "53076" "53077"
#>  [451] "53078" "53079" "53080" "53082" "53083" "53084" "53085" "53086" "53087"
#>  [460] "53088" "53089" "53090" "53091" "53093" "53094" "53096" "53097" "53098"
#>  [469] "53099" "53100" "53101" "53102" "53103" "53104" "53105" "53106" "53107"
#>  [478] "53108" "53109" "53110" "53111" "53112" "53113" "53114" "53115" "53116"
#>  [487] "53117" "53118" "53119" "53120" "53121" "53122" "53123" "53124" "53125"
#>  [496] "53126" "53127" "53128" "53129" "53130" "53131" "53132" "53133" "53134"
#>  [505] "53135" "53136" "53137" "53139" "53140" "53141" "53142" "53143" "53144"
#>  [514] "53145" "53146" "53147" "53148" "53150" "53151" "53152" "53153" "53154"
#>  [523] "53155" "53156" "53157" "53158" "53160" "53161" "53162" "53163" "53164"
#>  [532] "53165" "53168" "53169" "53170" "53172" "53173" "53174" "53175" "53176"
#>  [541] "53177" "53178" "53179" "53180" "53181" "53182" "53184" "53185" "53186"
#>  [550] "53187" "53188" "53189" "53190" "53191" "53192" "53193" "53195" "53196"
#>  [559] "53197" "53198" "53199" "53200" "53201" "53202" "53203" "53204" "53206"
#>  [568] "53208" "53209" "53210" "53211" "53212" "53213" "53214" "53216" "53218"
#>  [577] "53219" "53220" "53221" "53222" "53223" "53224" "53225" "53226" "53228"
#>  [586] "53229" "53230" "53232" "53233" "53234" "53235" "53236" "53237" "53238"
#>  [595] "53240" "53242" "53243" "53245" "53246" "53247" "53248" "53249" "53250"
#>  [604] "53251" "53253" "53255" "53256" "53257" "53258" "53259" "53260" "53261"
#>  [613] "53262" "53263" "53264" "53265" "53266" "53267" "53269" "53270" "53271"
#>  [622] "53272" "53273" "53276" "72001" "72002" "72003" "72004" "72005" "72006"
#>  [631] "72007" "72008" "72009" "72010" "72011" "72012" "72013" "72015" "72016"
#>  [640] "72017" "72018" "72019" "72020" "72021" "72022" "72023" "72024" "72025"
#>  [649] "72026" "72027" "72028" "72029" "72031" "72032" "72034" "72035" "72036"
#>  [658] "72037" "72038" "72039" "72040" "72041" "72042" "72043" "72044" "72045"
#>  [667] "72046" "72047" "72048" "72049" "72050" "72051" "72052" "72053" "72054"
#>  [676] "72056" "72057" "72058" "72059" "72060" "72061" "72062" "72064" "72065"
#>  [685] "72066" "72067" "72068" "72070" "72071" "72072" "72073" "72074" "72075"
#>  [694] "72076" "72077" "72078" "72079" "72080" "72083" "72084" "72085" "72086"
#>  [703] "72087" "72088" "72089" "72090" "72091" "72093" "72094" "72095" "72096"
#>  [712] "72098" "72099" "72100" "72101" "72102" "72103" "72104" "72105" "72106"
#>  [721] "72107" "72109" "72110" "72111" "72112" "72113" "72114" "72115" "72118"
#>  [730] "72119" "72120" "72121" "72122" "72123" "72124" "72125" "72126" "72127"
#>  [739] "72128" "72129" "72130" "72131" "72132" "72133" "72134" "72135" "72136"
#>  [748] "72137" "72138" "72139" "72141" "72142" "72143" "72144" "72145" "72146"
#>  [757] "72147" "72148" "72149" "72150" "72151" "72152" "72153" "72154" "72155"
#>  [766] "72156" "72157" "72158" "72160" "72161" "72163" "72164" "72165" "72166"
#>  [775] "72167" "72168" "72169" "72170" "72171" "72172" "72173" "72174" "72175"
#>  [784] "72176" "72177" "72178" "72179" "72180" "72181" "72182" "72183" "72184"
#>  [793] "72185" "72186" "72187" "72188" "72189" "72190" "72191" "72192" "72193"
#>  [802] "72194" "72195" "72196" "72197" "72198" "72199" "72200" "72201" "72202"
#>  [811] "72204" "72205" "72208" "72209" "72210" "72211" "72212" "72213" "72214"
#>  [820] "72215" "72216" "72217" "72218" "72219" "72220" "72221" "72222" "72223"
#>  [829] "72224" "72225" "72226" "72227" "72228" "72229" "72230" "72231" "72232"
#>  [838] "72233" "72234" "72235" "72236" "72237" "72238" "72239" "72241" "72243"
#>  [847] "72244" "72245" "72246" "72247" "72248" "72249" "72250" "72251" "72252"
#>  [856] "72253" "72254" "72255" "72256" "72257" "72259" "72260" "72261" "72262"
#>  [865] "72264" "72265" "72266" "72267" "72268" "72269" "72270" "72271" "72272"
#>  [874] "72273" "72274" "72275" "72276" "72277" "72278" "72279" "72280" "72281"
#>  [883] "72282" "72283" "72286" "72287" "72289" "72290" "72291" "72292" "72293"
#>  [892] "72294" "72295" "72296" "72297" "72298" "72299" "72300" "72302" "72303"
#>  [901] "72305" "72306" "72307" "72308" "72309" "72310" "72311" "72312" "72313"
#>  [910] "72314" "72315" "72316" "72317" "72319" "72320" "72321" "72322" "72323"
#>  [919] "72324" "72325" "72326" "72327" "72328" "72329" "72330" "72331" "72332"
#>  [928] "72333" "72334" "72335" "72336" "72337" "72338" "72339" "72340" "72341"
#>  [937] "72342" "72343" "72344" "72345" "72346" "72347" "72348" "72349" "72350"
#>  [946] "72351" "72352" "72353" "72354" "72355" "72356" "72357" "72358" "72359"
#>  [955] "72360" "72361" "72362" "72363" "72364" "72366" "72367" "72368" "72369"
#>  [964] "72370" "72372" "72373" "72374" "72375" "72376" "72377" "72378" "72379"
#>  [973] "72380" "72381" "72382" "72383" "72385" "72386" "85001" "85002" "85003"
#>  [982] "85004" "85005" "85006" "85008" "85009" "85010" "85011" "85012" "85013"
#>  [991] "85014" "85015" "85016" "85017" "85018" "85019" "85020" "85021" "85022"
#> [1000] "85023" "85024" "85025" "85026" "85028" "85029" "85031" "85033" "85034"
#> [1009] "85035" "85036" "85037" "85038" "85039" "85040" "85041" "85042" "85045"
#> [1018] "85046" "85047" "85049" "85050" "85051" "85053" "85054" "85055" "85056"
#> [1027] "85058" "85059" "85061" "85062" "85064" "85065" "85066" "85067" "85070"
#> [1036] "85071" "85072" "85073" "85074" "85076" "85077" "85078" "85080" "85081"
#> [1045] "85082" "85083" "85084" "85086" "85087" "85088" "85089" "85090" "85092"
#> [1054] "85093" "85094" "85095" "85096" "85097" "85098" "85099" "85100" "85101"
#> [1063] "85102" "85103" "85104" "85105" "85106" "85108" "85109" "85110" "85111"
#> [1072] "85112" "85113" "85114" "85115" "85116" "85117" "85118" "85119" "85120"
#> [1081] "85121" "85123" "85125" "85126" "85127" "85128" "85129" "85130" "85131"
#> [1090] "85132" "85133" "85134" "85135" "85136" "85137" "85138" "85139" "85140"
#> [1099] "85141" "85142" "85143" "85144" "85145" "85146" "85147" "85148" "85149"
#> [1108] "85151" "85152" "85153" "85154" "85155" "85156" "85157" "85158" "85159"
#> [1117] "85160" "85161" "85162" "85163" "85164" "85167" "85169" "85171" "85172"
#> [1126] "85174" "85175" "85176" "85177" "85178" "85179" "85181" "85182" "85184"
#> [1135] "85185" "85186" "85187" "85188" "85189" "85190" "85191" "85192" "85193"
#> [1144] "85194" "85196" "85197" "85198" "85199" "85200" "85201" "85202" "85204"
#> [1153] "85205" "85206" "85207" "85208" "85209" "85210" "85211" "85213" "85214"
#> [1162] "85215" "85216" "85218" "85220" "85221" "85222" "85223" "85226" "85227"
#> [1171] "85229" "85231" "85232" "85233" "85234" "85235" "85236" "85237" "85238"
#> [1180] "85239" "85240" "85242" "85243" "85244" "85245" "85246" "85247" "85248"
#> [1189] "85250" "85251" "85252" "85254" "85255" "85256" "85259" "85260" "85261"
#> [1198] "85262" "85264" "85265" "85266" "85267" "85268" "85269" "85271" "85273"
#> [1207] "85274" "85276" "85277" "85278" "85280" "85281" "85282" "85284" "85285"
#> [1216] "85286" "85287" "85288" "85289" "85290" "85291" "85292" "85293" "85294"
#> [1225] "85295" "85296" "85297" "85298" "85300" "85301" "85302" "85303" "85304"
#> [1234] "85305" "85306"